[Ingestion Client] Add durationInSeconds and silence between segments to database, remove BOMs (#1260)
* Set up CI with Azure Pipelines

[skip ci]

* add text analytics nuget

* retry on 429

* remove min precision

* update nugets

* update function to be long running

* Update TextAnalyticsProvider.cs

* Update TextAnalyticsProvider.cs

* add db fields for durationInSeconds and silence between segments

* remove BOM

* update paths

+++ samples/ingestion/ingestion-client/Connector/Serializable/TranscriptionResult/RecognizedPhrase.cs
@@ -10,7 +10,7 @@ namespace Connector
 
     public class RecognizedPhrase
     {
-        public RecognizedPhrase(string recognitionStatus, int channel, int speaker, string offset, string duration, double offsetInTicks, double durationInTicks, IEnumerable<NBest> nBest)
+        public RecognizedPhrase(string recognitionStatus, int channel, int speaker, string offset, string duration, long offsetInTicks, long durationInTicks, IEnumerable<NBest> nBest)
         {
             RecognitionStatus = recognitionStatus;
             Channel = channel;
@@ -38,10 +38,10 @@ namespace Connector
         public string Duration { get; }
 
         [JsonProperty("offsetInTicks")]
-        public double OffsetInTicks { get; }
+        public long OffsetInTicks { get; }
 
         [JsonProperty("durationInTicks")]
-        public double DurationInTicks { get; }
+        public long DurationInTicks { get; }
 
         [JsonProperty("nBest")]
         public IEnumerable<NBest> NBest { get; set; }
